Added map locations and images
Added more map locations and images to go with them . There is a final boss area and a treasure chest to search.

+++ Adventure_Game_Final/map.cpp
@@ -10,7 +10,7 @@ Map::Map(){
 // Returns a game area from an ID
 // Returns an area with an ID of 0 if no area is found
 Area Map::getAreaById(std::string id){
-    for (int i = 0; i < mapData.size(); i++){
+    for (size_t i = 0; i < mapData.size(); i++){
         if (mapData[i].getId() == id){
             return mapData[i];
         }
@@ -109,7 +109,7 @@ void Map::loadAreasFromFile(std::string fileName){
 };
 
 void Map::addSearchItems(){
-    for (int i = 0; i < mapData.size(); i++){
+    for (size_t i = 0; i < mapData.size(); i++){
         if(mapData[i].getId() == "4"){
             mapData[i].setSearchItem(Item("Old key", "A black key with an ornate design", "noImage", false, 0));
         }
